WIP Rust device in/out refactoring

+++ rust/src/ixgbe/device_limits.rs
@@ -1,17 +1,17 @@
-pub const FIVE_TUPLE_FILTERS_COUNT: u32 = 128;
+pub const FIVE_TUPLE_FILTERS_COUNT: usize = 128;
 
-pub const INTERRUPT_REGISTERS_COUNT: u32 = 3;
+pub const INTERRUPT_REGISTERS_COUNT: usize = 3;
 
-pub const MULTICAST_TABLE_ARRAY_SIZE: u32 = 4 * 1024;
+pub const MULTICAST_TABLE_ARRAY_SIZE: usize = 4 * 1024;
 
-pub const POOLS_COUNT: u32 = 64;
+pub const POOLS_COUNT: usize = 64;
 
-pub const RECEIVE_ADDRESSES_COUNT: u32 = 128;
+pub const RECEIVE_ADDRESSES_COUNT: usize = 128;
 
-pub const RECEIVE_QUEUES_COUNT: u32 = 128;
+pub const RECEIVE_QUEUES_COUNT: usize = 128;
 
-pub const TRANSMIT_QUEUES_COUNT: u32 = 128;
+pub const TRANSMIT_QUEUES_COUNT: usize = 128;
 
-pub const TRAFFIC_CLASSES_COUNT: u32 = 8;
+pub const TRAFFIC_CLASSES_COUNT: usize = 8;
 
-pub const UNICAST_TABLE_ARRAY_SIZE: u32 = 4 * 1024;
+pub const UNICAST_TABLE_ARRAY_SIZE: usize = 4 * 1024;
